+++ Twig/EventsExtension.php
@@ -1,0 +1,46 @@
+<?php
+
+/*
+ * This file is part of the AdminLTE bundle.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace KevinPapst\AdminLTEBundle\Twig;
+
+use KevinPapst\AdminLTEBundle\Event\SidebarMenuEvent;
+use KevinPapst\AdminLTEBundle\Event\SidebarUserEvent;
+use KevinPapst\AdminLTEBundle\Model\UserInterface;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
+use Symfony\Component\HttpFoundation\Request;
+use Twig\Extension\RuntimeExtensionInterface;
+
+final class EventsExtension implements RuntimeExtensionInterface
+{
+    /**
+     * @var EventDispatcherInterface
+     */
+    private $eventDispatcher;
+
+    public function __construct(EventDispatcherInterface $dispatcher)
+    {
+        $this->eventDispatcher = $dispatcher;
+    }
+
+    public function getMenu(Request $request): array
+    {
+        /** @var SidebarMenuEvent $event */
+        $event = $this->eventDispatcher->dispatch(new SidebarMenuEvent($request));
+
+        return $event->getItems();
+    }
+
+    public function getSidebarUser(): ?UserInterface
+    {
+        /** @var SidebarUserEvent $event */
+        $event = $this->eventDispatcher->dispatch(new SidebarUserEvent());
+
+        return $event->getUser();
+    }
+}
